+++ source/platforms/bitbucket_server/BitBucketServerGit.ts
@@ -12,7 +12,6 @@ import { BitBucketServerAPI } from "../bitbucket_server/BitBucketServerAPI"
 import { GitJSONToGitDSLConfig, gitJSONToGitDSL, GitStructuredDiff } from "../git/gitJSONToGitDSL"
 
 import * as debug from "debug"
-import { EOL } from "os"
 const d = debug("danger:BitBucketServerGit")
 
 /**
@@ -37,15 +36,15 @@ function bitBucketServerCommitToGitCommit(
     },
     committer: bbsCommit.committer
       ? {
-        email: bbsCommit.committer.emailAddress,
-        name: bbsCommit.committer.name,
-        date: new Date(bbsCommit.committerTimestamp).toISOString(),
-      }
+          email: bbsCommit.committer.emailAddress,
+          name: bbsCommit.committer.name,
+          date: new Date(bbsCommit.committerTimestamp).toISOString(),
+        }
       : {
-        email: bbsCommit.author.emailAddress,
-        name: bbsCommit.author.name,
-        date: new Date(bbsCommit.authorTimestamp).toISOString(),
-      },
+          email: bbsCommit.author.emailAddress,
+          name: bbsCommit.author.name,
+          date: new Date(bbsCommit.authorTimestamp).toISOString(),
+        },
     message: bbsCommit.message,
     tree: null,
     url,
@@ -98,14 +97,24 @@ const bitBucketServerDiffToGitJSONDSL = (diffs: BitBucketServerDiff[], commits: 
 }
 
 const bitBucketServerDiffToGitStructuredDiff = (diffs: BitBucketServerDiff[]): GitStructuredDiff => {
+  // We need all changed lines with it's type. It will convert hunk segment lines to flatten changed lines.
+  const segmentValues = { ADDED: "add", CONTEXT: "normal", REMOVED: "del" }
   return diffs.map(diff => ({
     from: diff.source && diff.source.toString,
     to: diff.destination && diff.destination.toString,
-    chunks: diff.hunks && diff.hunks.map(hunk => ({
-      changes: hunk.segments.map(segment => ({
-        type: segment.type === "ADDED" ? ("add" as "add") : ("del" as "del"),
-        content: segment.lines.map(({ line }) => line).join(EOL),
+    chunks:
+      diff.hunks &&
+      diff.hunks.map(hunk => ({
+        changes: hunk.segments
+          .map(segment =>
+            segment.lines.map(line => ({
+              type: segmentValues[segment.type] as "add" | "del" | "normal",
+              content: line.line,
+              sourceLine: line.source,
+              destinationLine: line.destination,
+            }))
+          )
+          .reduce((a, b) => a.concat(b), []),
       })),
-    })),
   }))
 }