+++ source/runner/templates/_tests/_bitbucketServerTemplate.test.ts
@@ -43,28 +43,28 @@ describe("generating messages for BitBucket server", () => {
 
 describe("generating inline messages", () => {
   it("Shows the failing message", () => {
-    const issues = inlineTemplate(failsResults)
-    expect(issues).toContain("- :no_entry_sign: Failing message")
-    expect(issues).not.toContain("- :warning:")
-    expect(issues).not.toContain("- :book:")
+    const issues = inlineTemplate("blankID", failsResults, "File.swift", 5)
+    expect(issues).toContain("- 🚫 Failing message")
+    expect(issues).not.toContain("- ⚠️")
+    expect(issues).not.toContain("- 📖")
   })
 
   it("Shows the warning message", () => {
-    const issues = inlineTemplate(warnResults)
-    expect(issues).toContain("- :warning: Warning message")
-    expect(issues).not.toContain("- :no_entry_sign:")
-    expect(issues).not.toContain("- :book:")
+    const issues = inlineTemplate("blankID", warnResults, "File.swift", 5)
+    expect(issues).toContain("- ⚠️ Warning message")
+    expect(issues).not.toContain("- 🚫")
+    expect(issues).not.toContain("- 📖")
   })
 
   it("Shows the message", () => {
-    const issues = inlineTemplate(messagesResults)
-    expect(issues).toContain("- :book: Message")
-    expect(issues).not.toContain("- :no_entry_sign:")
-    expect(issues).not.toContain("- :warning:")
+    const issues = inlineTemplate("blankID", messagesResults, "File.swift", 5)
+    expect(issues).toContain("- 📖 Message")
+    expect(issues).not.toContain("- 🚫")
+    expect(issues).not.toContain("- ⚠️")
   })
 
   it("Shows markdowns one after another", () => {
-    const issues = inlineTemplate(markdownResults)
+    const issues = inlineTemplate("blankID", markdownResults, "File.swift", 5)
     const expected = `
 ### Short Markdown Message1