+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 * Fixed a bug where Danger posts empty main comment when it have one or more inline comments to post [@codestergit][]
 * fix bug when commiting .png files on BitBucket [@Mifi][]
+* Adds support for inline comments for bitbucket server. [@codestergit][]
 
 ## 3.4.7
 
@@ -1049,3 +1050,5 @@ Not usable for others, only stubs of classes etc. - [@orta][]
 [@mxstbr]: https://github.com/mxstbr
 [@happylinks]: https://github.com/happylinks
 [@fwal]: https://github.com/fwal
+[@codestergit]: https://github.com/codestergit
+